set default device type, refactor changes
