+++ internal/db/devices_test.go
@@ -5,14 +5,11 @@ import (
 	"database/sql"
 	"fmt"
 	"log"
-	"os"
-	"strconv"
 	"testing"
 	"time"
 
 	"github.com/checkspeed/sc-backend/internal/models"
 	"github.com/google/uuid"
-	"github.com/joho/godotenv"
 	_ "github.com/lib/pq"
 	"github.com/ory/dockertest/v3"
 	"github.com/ory/dockertest/v3/docker"
@@ -171,125 +168,17 @@ func Test_GetIDByIdentifier(t *testing.T) {
 	})
 }
 
-func Test_RunManualUpMigration(t *testing.T) {
-	ctx := context.Background()
-	err := godotenv.Load("../../.env")
-	require.NoError(t, err)
-	databaseUrl := os.Getenv("MIGRATE_DB_URL")
-	store, err := NewStore(databaseUrl)
-	require.NoError(t, err)
-	t.Run("up migration", func(t *testing.T) {
-		migrator, err := NewMigrator(store)
-		require.NoError(t, err)
-		err = migrator.Up(ctx)
-		assert.NoError(t, err)
-	})
-}
-
-func TestXxx(t *testing.T) {
-	err := godotenv.Load("../../.env")
-	require.NoError(t, err)
-	databaseUrl := os.Getenv("OLD_DB_URL")
-	store, err := NewStore(databaseUrl)
-	require.NoError(t, err)
-
-	var speedTestResult []models.SpeedtestResultsOld
-	result := store.db.Table("speed_test_results").Find(&speedTestResult)
-	require.NoError(t, result.Error)
-
-	fmt.Println(len(speedTestResult))
-	fmt.Println(speedTestResult[23].Latitude, ReduceFloat(speedTestResult[23].Latitude), ReduceFloat(speedTestResult[23].Latitude) == 6.5)
-	fmt.Println(speedTestResult[45].Latitude, ReduceFloat(speedTestResult[45].Latitude) == 6.4)
-
-	// Victor details
-	deviceID1 := ""
-	deviceID2 := ""
-
-	var usCount, kCount, vCount, tCount int
-	var newData []models.SpeedTestResults
-	for _, input := range speedTestResult {
-		if input.ISP == "test" {
-			tCount++
-			continue
-		}
-
-		deviceID := deviceID2
-		kCount++
-		if ReduceFloat(input.Latitude) == 6.5 {
-			kCount--
-			vCount++
-			deviceID = deviceID1
-		}
-
-		state := "Lagos"
-		countryCode := "NG"
-		CountryName := "Nigeria"
-		ContinentCode := "AF"
-		ContinentName := "Africa"
-
-		if input.ISP == "Google LLC" || input.ISP == "Google" {
-			state = "Florida"
-			countryCode = "US"
-			CountryName = "United States of America"
-			ContinentCode = "NA"
-			ContinentName = "North America"
-
-			usCount++
-		}
-
-		s := models.SpeedTestResults{
-			ID:               uuid.NewString(),
-			DownloadSpeed:    input.DownloadSpeed,
-			MaxDownloadSpeed: input.MaxDownloadSPeed,
-			MinDownloadSpeed: input.MinDownloadSpeed,
-			TotalDownload:    input.TotalDownload,
-			UploadSpeed:      input.UploadSpeed,
-			MaxUploadSpeed:   input.MaxUploadSpeed,
-			MinUploadSpeed:   input.MinUploadSpeed,
-			TotalUpload:      input.TotalUpload,
-			Latency:          input.Latency,
-			LoadedLatency:    input.LoadedLatency,
-			UnloadedLatency:  input.UnloadedLatency,
-			DownloadLatency:  input.DownloadLatency,
-			UploadLatency:    input.UploadLatency,
-			DeviceID:         deviceID,
-			ISP:              input.ISP,
-			ISPCode:          input.ISPCode,
-			ConnectionType:   input.ConnectionType,
-			ConnectionDevice: input.ConnectionDevice,
-			TestPlatform:     input.TestPlatform,
-			// ServerID:         input.TestServer.ID,
-			ServerName:     input.ServerName,
-			State:          state,
-			CountryCode:    countryCode,
-			CountryName:    CountryName,
-			ContinentCode:  ContinentCode,
-			ContinentName:  ContinentName,
-			Longitude:      input.Longitude,
-			Latitude:       input.Latitude,
-			LocationAccess: input.LocationAccess,
-			TestTime:       input.TestTime,
-			CreatedAt:      input.CreatedAt,
-			UpdatedAt:      time.Now(),
-		}
-
-		newData = append(newData, s)
-	}
-
-	fmt.Println(len(newData), usCount, vCount, kCount, tCount)
-	fmt.Println(newData[0])
-	// databaseUrl = os.Getenv("DB_URL")
-	// newStore, err := NewStore(databaseUrl)
-	// require.NoError(t, err)
-
-	// res := newStore.db.Create(&newData)
-	// require.NoError(t, res.Error)
-
-}
-
-func ReduceFloat(value float64) float64 {
-	// Format the float to 1 decimal place and then parse it back to a float
-	formattedValue := fmt.Sprintf("%.1f", value)
-	reducedValue, _ := strconv.ParseFloat(formattedValue, 64)
-	return reducedValue
-}
+// func Test_RunManualUpMigration(t *testing.T) {
+// 	ctx := context.Background()
+// 	err := godotenv.Load("../../.env")
+// 	require.NoError(t, err)
+// 	databaseUrl := os.Getenv("MIGRATE_DB_URL")
+// 	store, err := NewStore(databaseUrl)
+// 	require.NoError(t, err)
+// 	t.Run("up migration", func(t *testing.T) {
+// 		migrator, err := NewMigrator(store)
+// 		require.NoError(t, err)
+// 		err = migrator.Up(ctx)
+// 		assert.NoError(t, err)
+// 	})
+// }
